Task 5
Write a Python script that calculates the distance between any two points on the Earth’s surface, given their latitude and longitude.

The script should first prompt the user to enter the coordinates of each location, then calculate the spherical distance along the Earth’s surface (use the following distance formula). If the user does not enter numbers for latitude or longitude values, prompt them to re-enter.

To make the problem simpler, you may assume that numbers entered are valid latitude or longitude coordinates in decimal degrees, but feel free to experiment with responding to violations of this assumption.

Finally, save your script as great_circle_dist.py. (8 pts)
Spherical distance formula:

screenshot of Spherical distance formula

whereby d is the angle between two points, λ1 and λ2 are the longitudes of the points, and Φ1 and Φ2 are the latitudes of the points. To convert the angle to a spherical distance, multiply by the radius of the earth, which is ~6300 km.